pre-processed up to binnification
wrote the pipeline to preprocess and quick vis. for most analysis we need same timebin numbers, so using visualisation, I subset first he emotion expressions (leaving out the spoken for separate processing). And chose a bin number based on visualisation that would preserve much of the original sampling while covering all ranges and preserving [patterns. (e.g. on smoothed data). those was a 100bins where we average the AU weights. this means that for every bin we have a minimum of two samples