visualised NMF for k4 and k5
this showed some differentiation between drug and placebo for anger only, but nothing pronounced, so whatever effects are  here are likely minor.

k3 and k5=4 results are similar, next I will try and estimate NMF separately, on drug and placebo data. While I don't think it should change it could be that the the solution when pooled is toward a global optimum rather than a local optimum