temporal k-mean on diff and speed
one possibility to deal with the temporal structure is to conduct the mean clustering on difference scores and speed smothed with moving average. this seems to return intuitive results fro k2 and k = 3, namely the clusters capture sustainment, activation, or sustainment, activation, sustain, deactivation.

we can compare the time of activation and sustain, but also the values, for example does drig and placebo differ on the means for each cluster