temporal kmean on diff
this approach is able to pick up clear transition states in some cases but not so clear in other cases. should try a different version with just speed.
also need to figure out what to do with very small clusters and ordering or naming the clusters according to rule that can look back to diff an speed , e.g. if diff is lower than average probably sustain