fixed panel and compared ggside vs ggmarginal plotting
fixed the issues with distorted plots in patchwork panels. Solution is to wrap plots before patchworking it.

There is still a a mall issue where I can't set a fixed order for the side boxplots. ggside should do this but in my data, it won't show the labels, whereas i can see the labels in a mock example with iris dataset